Interfaz ahora corre en un hilo distinto. Las actualizaciones todavía no corresponden al progreso real. El código que se tiene es de prueba.

+++ ProgressbarList.h
@@ -4,7 +4,7 @@
 #define PROGRESSBARLIST
 
 void progressbarlist_init(GtkBuilder* builder, int length);
-void progressbarlist_item_update(long id, double result, double percentage, int tickets);
+void progressbarlist_item_update(int id, double result, double percentage, int tickets);
 
 GObject* progressbarlist_get_label(int id);
 GObject* progressbarlist_get_progressbar(int id);
